code folder structure reorganization
