+++ phase_2/data_processing/experimental/fit_plane_to_depth_map.py
@@ -5,9 +5,9 @@ from scipy.linalg import svd
 import pandas as pd
 import matplotlib.pyplot as plt
 from pypfm import PFMLoader
-from get_road_boundary_points import get_image_road_points
-from utils import get_depth_data, get_depth_of_pixel
-from align_segmented_road_with_lidar import transform_2d_points_to_3d, INIT_CAMERA_PARAMS, \
+from data_processing.get_road_boundary_points import get_image_road_points
+from data_processing.utils import get_depth_data, get_depth_of_pixel
+from data_processing.align_segmented_road_with_lidar import transform_2d_points_to_3d, INIT_CAMERA_PARAMS, \
     FOCAL_LENGTH_X, FOCAL_LENGTH_Y
 
 DEPTH_SCALING_FACTOR = 189
